Minor fix for user preference "full screen" getting reset
User reported that when using MS mod, their preference for "full screen" was lost. Disabling the mod would result in the game starting in wIndowed mode.

Fix: store player's preference prior to activating additional displays, wait for main menu to load, then allow a few seconds for the displays to "settle". Lastly, set the main display to full screen mode if that was the user's preference.